microservicio_auditoria/requirements.txt
Flask==2.3.2
pymongo==4.4.0
python-dotenv==1.0.0
requests==2.31.0
